Extract StoredMessageSorter as outer class
